x86: request higher HWP state
This change sets a balanced performance setting which leans more towards
high performance. The desired performance is requested to be the
maximum. The minimum requested performance is set to the guaranteed
performance level.

Note that the guaranteed performance is only valid during the current
interval. As the kernel never updates this, the value that is announced
during CPU boot-up will stay forever.

The Intel manual states that the HWP capabilities are updated every ~1
second.

Change-Id: If85e9f0bc8fa01b908297f4105a3a444ca9531a1
Signed-off-by: Matthias Lange <matthias.lange@kernkonzept.com>