Abort remote IPC send DRQ when send timeout expires
IPC timeouts on the sender side did not expire when the receiver is on a
remote CPU (cross-core IPC) and is blocked by a higher priority context.
The root cause of this issue is that we implement cross-core IPC with a
DRQ executed by the sender on the receiver, which until now could not be
aborted prematurely. So the sender had to wait until the receiver gets
scheduled, handles the DRQ and sends an answer DRQ.

To take back control, i.e. to enforce the IPC send timeout it specified,
the sender now programs an IPC_remote_timeout. When the timeout expires,
it tries to abort the DRQ, i.e. dequeue the DRQ from the receiver's DRQ
queue, which succeeds in case the receiver has not yet executed the DRQ.

This mechanism does not apply to zero timeouts, for those there will be
a solution in a change on top.

See ticket #CD-536.

Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: Idce95f6907020ca44551f96fdfc81ef5f841eb94
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>