thread: Abort pending migration in force_to_invalid_cpu()
After forcing a thread to the invalid CPU, we must also abort a
migration potentially pending for the thread. This is necessary,
because otherwise the migration will never be executed, since we just
moved the thread to the invalid CPU and removed its _pending_rq,
so Context::Pending_rqq::handle_requests() will never notice the
pending migration, making the initiator wait forever.

Change-Id: I45f74699014134c47168622558fe3f3c38cf5e11
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>