[API] Execute remote IPC send DRQs eagerly
Use the newly introduced eager DRQ mechanism to implement proper zero
timeout semantics for cross-core IPC operations.

The sender CPU determines with a heuristic check if the receiver thread
is ready to receive from the sender. If the check indicates that the
receiver might be ready-to-receive (since it is racy we cannot get more
than a might) the sender performs the remote IPC send as an eager DRQ.
To ensure consistent behavior, we do this regardless of the timeout.

Only if the check indicates that the receiver is not ready, the
procedure varies depending on the timeout:
  - Zero: Immediately return a timeout error. This may be racy, but that
          same race would happen anyways (at a different time) and the
          zero-timeout IPC would fail.
  - Finite: Program IPC_remote_timeout, which aborts DRQ once send
            timeout expires, and perform the remote IPC send as a
            regular DRQ.
  - Infinite: Perform the remote IPC send as a regular DRQ.

There is a special case for IPC replies, which are always sent eagerly
(necessary because of timing issue, where the xcpu_state_change() on the
caller was not yet applied or is not yet visible to the callee).

This timing issue is also the reason why the change is considered
API-breaking. Previously it was possible to reply to an IPC call with a
non-reply IPC send operation targeting the caller. But now with eager
DRQ execution, using a send with a zero timeout to reply to a cross-core
IPC call is racey, because the heuristic check can determine that the
partner is not receive ready, resulting in a send timeout error.

Eager DRQ execution may allow senders to threads on the receiver CPU
(even to low prio threads), to interfere with the execution timing on
the receiver CPU (effectively stealing CPU time through the eager DRQ
execution). However, since eager DRQs are only sent to receive-ready
threads, the impact is limited.

See ticket #CD-536.

Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: If16999ba251abee8239df48f6abd1ec67641821e
Requires-Change-Id: I65c04bf362f15e811deb6292443d58a75ba79c2c
Requires-Change-Id: I53b78bf2b8630478ca1dbd14a6de03c8e8723854
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>